Gets metadata from specified source.

Gets metadata from specified source.

get_metadata(
  sourcetype,
  from_dir = getwd(),
  filename = "textot.txt",
  encoding = "UTF-8",
  rtibble
)

Arguments

sourcetype

the type of source ("rtibble","txm" or "iramuteq")

from_dir

the corpus directory (contains metadata.csv and multiple .txt files)

filename

the name of the Iramuteq-formatted txt file (defaults to "textot.txt")

encoding

the encoding of the TXM corpus (defaults to UTF-8)

rtibble

the name of the R tibble, if relevant (sourcetype="rtibble")

Value

a tibble of metadata

Examples

get_metadata(sourcetype="txm",
             from_dir=system.file("extdata/fables",package="textoteR"),
             encoding="UTF-8")
#> # A tibble: 9 × 3
#>   id    titre                                                      date
#>   <chr> <chr>                                                     <int>
#> 1 fab01 Le lièvre et la Tortue                                     1668
#> 2 fab02 Le Corbeau et le Renard                                    1668
#> 3 fab03 Le Cochet, le Chat et le Souriceau                         1668
#> 4 fab04 Le Renard et les Raisins                                   1668
#> 5 fab05 Le Rat de villes et le Rat des champs                      1668
#> 6 fab06 Le Chêne et le Roseau                                      1668
#> 7 fab07 La Cigale et la Fourmi                                     1668
#> 8 fab08 La Grenouille qui se veut faire aussi grosse que le Boeuf  1668
#> 9 fab09 Le Loup et l’Agneau                                        1668
get_metadata(sourcetype="iramuteq",
             from_dir=system.file("extdata",package="textoteR"),
             filename="macron_covid.txt",
             encoding="UTF-8")
#> # A tibble: 5 × 2
#>   id    date    
#>   <chr> <chr>   
#> 1 txt1  24112020
#> 2 txt2  28102020
#> 3 txt3  14062020
#> 4 txt4  13042020
#> 5 txt5  16032020
get_metadata(sourcetype="rtibble",
             rtibble=LVtweets)
#> # A tibble: 244 × 5
#>    id     created_at          is_quote is_retweet favorite_count
#>    <chr>  <dttm>              <lgl>    <lgl>               <dbl>
#>  1 txt001 2020-11-09 13:14:11 FALSE    FALSE                   9
#>  2 txt002 2020-10-23 08:50:15 FALSE    FALSE                   1
#>  3 txt003 2020-10-23 08:40:41 FALSE    FALSE                   2
#>  4 txt004 2020-10-07 09:26:55 FALSE    FALSE                  12
#>  5 txt005 2020-06-12 06:46:58 FALSE    FALSE                   0
#>  6 txt006 2020-06-12 06:42:41 FALSE    FALSE                   1
#>  7 txt007 2020-06-11 16:39:37 FALSE    FALSE                  15
#>  8 txt008 2020-06-03 08:02:12 FALSE    FALSE                   1
#>  9 txt009 2020-06-03 07:55:05 FALSE    FALSE                   0
#> 10 txt010 2020-06-03 07:35:59 FALSE    FALSE                   0
#> # ℹ 234 more rows