Gets text data from specified source.

Gets text data from specified source.

get_textdata(
  sourcetype,
  from_dir = getwd(),
  filename = "textot.txt",
  encoding = "UTF-8",
  rtibble
)

Arguments

sourcetype

the type of source ("rtibble","txm" or "iramuteq")

from_dir

the corpus directory (contains metadata.csv and multiple .txt files)

filename

the name of the Iramuteq-formatted txt file (defaults to "textot.txt")

encoding

the encoding of the TXM corpus (defaults to UTF-8)

rtibble

the name of the R tibble, if relevant (sourcetype="rtibble")

Value

a tibble of metadata

Examples

get_textdata(sourcetype="txm",
             from_dir=system.file("extdata/fables",package="textoteR"),
             encoding="UTF-8")
#> # A tibble: 9 × 2
#>   id    text                                                                    
#>   <chr> <chr>                                                                   
#> 1 fab01 "Rien ne sert de courir ; il faut partir à point. Le Lièvre et la Tortu…
#> 2 fab02 "Maître corbeau, sur un arbre perché, Tenait en son bec un fromage. Maî…
#> 3 fab03 "Un Souriceau tout jeune, et qui n'avait rien vu, Fut presque pris au d…
#> 4 fab04 "Certain Renard Gascon, d'autres disent Normand, Mourant presque de fai…
#> 5 fab05 "Autrefois le Rat de ville Invita le Rat des champs, D'une façon fort c…
#> 6 fab06 "Le Chêne un jour dit au Roseau : \"Vous avez bien sujet d'accuser la N…
#> 7 fab07 "La Cigale, ayant chanté Tout l'été, Se trouva fort dépourvue Quand la …
#> 8 fab08 "Une Grenouille vit un Boeuf Qui lui sembla de belle taille. Elle qui n…
#> 9 fab09 "La raison du plus fort est toujours la meilleure : Nous l'allons montr…
get_textdata(sourcetype="iramuteq",
             from_dir=system.file("extdata",package="textoteR"),
             filename="macron_covid.txt",
             encoding="UTF-8")
#> # A tibble: 5 × 2
#>   id    text                                                                    
#>   <chr> <chr>                                                                   
#> 1 txt1  Françaises, Français, Mes chers compatriotes, Quand, le 28 octobre, je …
#> 2 txt2  Françaises, Français, Mes chers compatriotes, La dernière fois que je m…
#> 3 txt3  Françaises, Français, Mes chers compatriotes, Je veux ce soir vous parl…
#> 4 txt4  Françaises, Français, Mes chers compatriotes, Nous sommes en train de v…
#> 5 txt5  Françaises, Français, Jeudi soir, je me suis adressé à vous pour évoque…
get_textdata(sourcetype="rtibble",
             rtibble=LVtweets)
#> # A tibble: 244 × 2
#>    id     text                                                                  
#>    <chr>  <chr>                                                                 
#>  1 txt001 "Nouveau billet de blog: #Géocoder sous #rstats via une #API https://…
#>  2 txt002 "@3wen @analysingstuffs Pourtant y'aurait moyen de faire des beaux de…
#>  3 txt003 "@analysingstuffs @3wen Haha arrêtez vous allez me faire rougir... Ja…
#>  4 txt004 "Nouveau billet de blog! Utiliser un package https://t.co/E5ApdBhnKK …
#>  5 txt005 "@datagistips :-) j'avoue la \"démo\" de flowrpowr faisait AUSSI part…
#>  6 txt006 "@fbahoken Tout le plaisir est pour moi :-)"                          
#>  7 txt007 "Nouveau billet: \"#Janitor: nettoie-moi cette table, et qu'elle bril…
#>  8 txt008 "@ma_salmon @StatnMap quite a programme!"                             
#>  9 txt009 "@ma_salmon @StatnMap Maëlle, that's intriguing, what is your course …
#> 10 txt010 "@ValdenaireEric @lblepage Merci! J'ai utilisé la solution \"Google A…
#> # ℹ 234 more rows